+++ sources/renderer/resource_loader.hpp
@@ -12,12 +12,12 @@ private:
     static CommandPool*   m_command_pool;
     static CommandBuffer* m_cmd;
 
-    static Buffer* create_staging_buffer(u32 size, const void* data);
+    static Buffer* create_staging_buffer(u64 size, const void* data);
 
 public:
     static void init(const Device* m_device);
     static void shutdown();
-    static void upload_buffer(Buffer* buffer, u32 offset, u32 size, const void* data);
-    static void upload_image(Image* image, u32 size, const void* data);
+    static void upload_buffer(Buffer* buffer, u64 offset, u64 size, const void* data);
+    static void upload_image(Image* image, u64 size, const void* data);
 };
 } // namespace fluent
